+++ algorithms/0106-construct-binary-tree-from-inorder-and-postorder-traversal.go
@@ -1,0 +1,21 @@
+package algorithms
+
+func buildTree(inorder []int, postorder []int) *TreeNode {
+	if len(inorder) == 0 {
+		return nil
+	}
+
+	root := &TreeNode{Val: postorder[len(postorder)-1]}
+
+	i := 0
+	for ; i < len(inorder); i++ {
+		if root.Val == inorder[i] {
+			break
+		}
+	}
+
+	root.Left = buildTree(inorder[:i], postorder[:i])
+	root.Right = buildTree(inorder[i+1:], postorder[i:len(postorder)-1])
+
+	return root
+}